plaidservice/requirements.txt
Jinja2>=2.9.5
plaid-python